==> 02_code/10_trend_analysis.Rmd <==
---
title: "10 - Trend analysis"
author: "Jens Linden"
output:
  html_document:
    number_sections: yes
    toc: yes
---

STATUS: Development

# Management summary

* Simple analysis to identify trend of reviews for each hotel over time
* Business idea: identify hotels with negative trend and provide alert to take actions
* Focus on 'long-term trends' within the 2-year observation period
* Modeling trend as slope of linear model of reviews over time


# Inits
Initialisations and loading of libraries required:
```{r, warning = FALSE, message = FALSE, results = FALSE}
rm(list=ls())   # clear all
graphics.off()  # close all
Sys.setenv(LANG = "en")
# Libs
# library(checkpoint) # Package management to avoid incompatibilities; not required when dockerized
# checkpoint('2017-10-01') # Use all packages as of date provided
library(data.table) # Efficient and fast data munging in R; just do it
options(datatable.auto.index=FALSE) # Bug fixing for v < 1.9.5
library(ggplot2) # Good quality plots
library(rprojroot) # Rroot folder management
ROOT <- find_root(has_file("PROJECT_ROOT_DIR"))
library(knitr) # Knitting R markdown reports
source(paste0(ROOT, "/02_code/lib.R"))
options(width = 100)
```

# Load data
```{r}
hot <- readRDS(paste0(ROOT, '/01_data/03_preprocessed/hotels.rds'))
rev <- readRDS(paste0(ROOT, '/01_data/03_preprocessed/reviews.rds'))
```

# Compute trend of review score with respect to time

## Inspect example
```{r}
# Inspect columns of interest
rev[, .(id_rev, id_hot, Reviewer_Score, Review_Date)][1:3]
# Inspect reviews for one random hotel
ggplot(data = rev[id_hot == 78],
       aes(x = Review_Date, y = Reviewer_Score)) +
  geom_point() + 
  geom_smooth(method = 'lm') + 
  ggtitle('Reviews for hotel with ID 78')
```

## Compute trend as slope of linear model
```{r}
# Compute linear model for each hotel
rev_trend <- rev[, 
             .(rev_trend = lm(as.formula("Reviewer_Score ~ Review_Date"), 
                      data=.SD, 
                      na.action=na.exclude)$coefficients[2]), 
             by=c("id_hot")]
# Inspect result
rev_trend[order(rev_trend)]
```

## Plausichecks
```{r}
# Plot most negative slope
ggplot(data = rev[id_hot == 762],
       aes(x = Review_Date, y = Reviewer_Score)) +
  geom_point() + 
  geom_smooth(method = 'lm') + 
  ggtitle('Reviews for hotel with ID 762')
# Plot most negative slope
ggplot(data = rev[id_hot == 1012],
       aes(x = Review_Date, y = Reviewer_Score)) +
  geom_point() + 
  geom_smooth(method = 'lm') + 
  ggtitle('Reviews for hotel with ID 1012')
# Looks plausible

```

# Join to hotel data
```{r}
hot <- merge(hot, rev_trend, by = 'id_hot')
```

# Inspect extreme slopes
```{r}
# Inspect extreme slope values
ggplot(hot) + 
  geom_point(aes(x = rev_trend, y = n_reviews)) + 
  ylim(0, 100) +
  geom_hline(yintercept = 30) + 
  ggtitle('Number of reviews vs. review trend')
# Moste extreme solopes occure due to few data points (high variance)
hot[n_reviews < 30, .N] # 72 hotels have less than 30 reviews; remove when trend is utilzed further
```

# Inspect distribution of slopes
```{r}
ggplot(hot) + 
  geom_density(aes(x = rev_trend))
quantile(hot$rev_trend, probs = c(0.25, 0.75))
```


# Save results to file
```{r}
saveRDS(rev_trend,  paste0(ROOT, '/01_data/04_results/10_rev_trend.rds'))
```
