==> 02_code/35_drivers_per_hotel.Rmd <==
---
title: "35 - Identify drivers to improve scoring per hotel"
author: "Jens Linden"
output:
  html_document:
    number_sections: yes
    toc: yes
---

STATUS: Development

# Management summary
Build regression trees (with target being review score) for each hotel individually. The tree can reveal insights, where the hotel can improve, in order to obtain better ratings.

# Inits
Initialisations and loading of libraries required:
```{r, warning = FALSE, message = FALSE, results = FALSE}
rm(list=ls())   # clear all
graphics.off()  # close all
Sys.setenv(LANG = "en")
# Libs
# library(checkpoint) # Package management to avoid incompatibilities; not required when dockerized
# checkpoint('2017-10-01') # Use all packages as of date provided
library(data.table) # Efficient and fast data munging in R; just do it
options(datatable.auto.index=FALSE) # Bug fixing for v < 1.9.5
library(ggplot2) # Good quality plots
library(rprojroot) # Rroot folder management
ROOT <- find_root(has_file("PROJECT_ROOT_DIR"))
library(caret) # Classification And REgression Training
source(paste0(ROOT, "/02_code/lib.R"))
options(width = 100)
```

# Load data
```{r}
# Analytics base table (ABT) for hotels
(abt_hot <- readRDS(paste0(ROOT, '/01_data/03_preprocessed/abt_hot.rds')))
# Hotel clusters from analysis 30
(clus_hot <- readRDS(paste0(ROOT, '/01_data/04_results/30_hotel_cluster.rds')))
# Analytics base table (ABT) for reviews
(abt_rev <- readRDS(paste0(ROOT, '/01_data/03_preprocessed/abt_rev.rds')))
# Review data
(rev <- readRDS(paste0(ROOT, '/01_data/03_preprocessed/reviews.rds')))
```

# Preprocessing

## Add cluster of hotel hotel ABT
```{r}
# Join on hotel ID
abt_hot <- merge(abt_hot,
                 clus_hot,
                 by = 'id_hot')
```

## Add variables to review ABT
```{r}
abt_rev <- merge(abt_rev,
                 rev[, .(id_rev, Total_Number_of_Reviews_Reviewer_Has_Given, Reviewer_Score)],
                 by = 'id_rev')
# Inspect
abt_rev[1:3]
```

## Convert variable types
```{r}
# Identify character columns
(cols_char <- setdiff(dt_classes(abt_rev)[class == 'character', col], 'id_rev'))
# Convert to factors (categorical variables)
abt_rev[, (cols_char) := lapply(.SD, as.factor), .SDcols = cols_char]
# Check if worked
dt_classes(abt_rev)
```


# Regression tree for single hotel
Use a regression tree to model the reviewer score for each hotel indiviudally. Using `ctree` implementation of `party` package, as this allows to specify a minimum p-value for each split. The regression tree allows to deduce insights, where a hotel could improve to increase the ratings hence customer satisfaction, which will influence future revenue.

## Settings for modeling
```{r}
s <- list()
# Define hotel of interest
s$id_hot <- 3
# Define target variable
s$target <- 'Reviewer_Score'
# Define features
(s$feat <- setdiff(names(abt_rev), c('id_rev', 'id_hot', s$target)))
```

## Check for zero variance predictors
```{r}
(idx_nzv <- nearZeroVar(abt_rev[id_hot == s$id_hot, s$feat, with = F]))
```

## Check for correltated predictors
```{r}
x <- model.matrix(
  as.formula(paste0(s$target, " ~ .")), 
  data = abt_rev[id_hot == s$id_hot, c(s$target, s$feat), with = F]
)[, -1]
(idx_cor <- findCorrelation(cor(x)))
```

## Check for missing data
```{r}
sum(!complete.cases(abt_rev[id_hot == s$id_hot, c(s$target, s$feat), with = F]))
```

## Build regression tree
```{r}
set.seed(123) # Random seed to create reproducible results
# Use resamling using bootstrap method to validate model
ctrl <- trainControl(method = "boot", 
                     number = 50)
grid <- expand.grid(.mincriterion = .95, # request a p-value of 0.05 for each split
                    .maxdepth =  as.integer(1:3))
fit <- train(
    x = abt_rev[id_hot == s$id_hot, s$feat, with = F],
    y = abt_rev[id_hot == s$id_hot, get(s$target)],
    method = "ctree2", # rpart, rpart1SE
    trControl = ctrl, 
    # controls = ctree_control(minbucket = 50),
    metric = "Rsquared", # Accuracy
    tuneGrid = grid) # set to NULL to swich off tuning grid
# Inspect model
print(fit)
resampleHist(fit)
# plot final model
plot(fit$finalModel, type = "simple")
```

## Conclusion
As expected the $R^2$ value for the regression tree is quite low. Nethertheless, as long as splits are statistically significant, the tree model can be used to deduce insights, where the hotel can improve. In this particular case for hotel #3, business travellers tend to give lower ratings (albeit overall still high with 8.9 on average). This information could be used to make the hotel more attractive for business travellers. Note, however, that the majority of reviewers are non-business travellers.

```{r}

```
