==> 02_code/36_hotel_leaders_laggers.Rmd <==
---
title: "36 - Identify leaders and laggers"
author: "Jens Linden"
output:
  html_document:
    number_sections: yes
    toc: yes
---

STATUS: Development

# Management summary
After clusters for hotels have been defined, declare leaders as overperformers being above the 75% quartile of mean score performance. Similarily, declare laggers as underperformers being below the 25% quartile of mean score performance. Save data to file.

# Inits
Initialisations and loading of libraries required:
```{r, warning = FALSE, message = FALSE, results = FALSE}
rm(list=ls())   # clear all
graphics.off()  # close all
Sys.setenv(LANG = "en")
# Libs
# library(checkpoint) # Package management to avoid incompatibilities; not required when dockerized
# checkpoint('2017-10-01') # Use all packages as of date provided
library(data.table) # Efficient and fast data munging in R; just do it
options(datatable.auto.index=FALSE) # Bug fixing for v < 1.9.5
library(ggplot2) # Good quality plots
library(rprojroot) # Rroot folder management
ROOT <- find_root(has_file("PROJECT_ROOT_DIR"))
source(paste0(ROOT, "/02_code/lib.R"))
options(width = 100)
```


# Load data
```{r}
# Hotel masater data
(hot <- readRDS(paste0(ROOT, '/01_data/03_preprocessed/hotels.rds')))
# Analytics base table (ABT) for hotels
(abt_hot <- readRDS(paste0(ROOT, '/01_data/03_preprocessed/abt_hot.rds')))
# Hotel clusters from analysis 30
(clus_hot <- readRDS(paste0(ROOT, '/01_data/04_results/30_hotel_cluster.rds')))
# Review data
(rev <- readRDS(paste0(ROOT, '/01_data/03_preprocessed/reviews.rds')))
```

# Preprocessing

## Add cluster of hotel review ABT
```{r}
# Join on hotel ID
rev <- merge(rev,
             clus_hot,
             by = 'id_hot', 
             all.x = T)
```

## Add cluster of hotel to hotel ABT
```{r}
# Join on hotel ID
abt_hot <- merge(abt_hot,
                 clus_hot,
                 by = 'id_hot')
```

# Compute mean score per cluster
```{r}
# Inspect overall mean
rev[, .(score_clus_mean = mean(Reviewer_Score))]
(score_clus_mean <- rev[, .(score_clus_mean = mean(Reviewer_Score)), by = 'clus'])
ggplot(score_clus_mean) + 
  geom_bar(aes(x = clus, y = score_clus_mean), stat = 'identity') +
  ggtitle('Mean score per hotel cluster')
# Add to ABT
abt_hot <- merge(abt_hot,
                 score_clus_mean,
                 by = 'clus')
```

# Compute mean score per hotel
```{r}
abt_hot <- merge(abt_hot,
                 rev[, .(score_hot_mean = mean(Reviewer_Score)), by = 'id_hot'],
                 by = 'id_hot')
# Inspect
print(abt_hot[1:3])
```


# Inspect Reviewer_Score distribution per cluster
```{r}
ggplot(rev) + 
  geom_histogram(aes(x = Reviewer_Score)) + facet_wrap('clus')
```

# Inspect mean hotel score per cluster
```{r}
ggplot(abt_hot) +
  geom_density(aes(x = score_hot_mean)) +
  ggtitle('Distribution of mean review score for all hotels')
ggplot(abt_hot) +
  geom_density(aes(x = score_hot_mean)) +
  facet_wrap('clus') +
  ggtitle('Distribution of mean review score for hotels in each cluster')
```

# Compute leaders and laggers
Computed as 25% and 75% quartile of score mean within cluster.
```{r}
# Compute quantiles
abt_hot[, score_q25 := quantile(score_hot_mean, probs = 0.25), by = 'clus']
abt_hot[, score_q75 := quantile(score_hot_mean, probs = 0.75), by = 'clus']
# Inspect
abt_hot[1:3]
```

Compute leaders and laggers
```{r}
abt_hot[score_hot_mean <= score_q25, lead_lag := "lagger"]
abt_hot[score_hot_mean >= score_q75, lead_lag := "leader"]
abt_hot[is.na(lead_lag), lead_lag := 'middle']
# Inspect
abt_hot[1:3]
abt_hot[, .N, by = lead_lag] # Looks fine
```

# Save data
```{r}
saveRDS(abt_hot, paste0(ROOT, '/01_data/04_results/36_abt_hot_incl_leaders_laggers.rds'))
```
